# Conflicts:
#	.github/workflows/azure-static-web-apps-icy-water-0ae577403.yml
#	.github/workflows/azure-static-web-apps-nice-beach-021ec5303.yml
#	.github/workflows/azure-static-web-apps-yellow-island-06ecb8603.yml
#	appPackage/appPackage.zip
#	appPackage/color.png
#	appPackage/manifest.json
#	npp/src/app/services/microsoft-data/teams.service.ts
#	npp/src/environments/environment.prod.ts
#	npp/src/environments/environment.ts

+++ npp/src/app/services/files.service.ts
@@ -13,9 +13,10 @@ export class FilesService {
 
   async uploadFileToFolder(fileData: string, folder: string, fileName: string, metadata?: any): Promise<any> {
     if (metadata) {
-      let scenarios = metadata.ModelScenarioId;
-      if (scenarios) {
-        let file = await this.getFileByScenarios(folder, scenarios);
+      const scenarios = metadata.ModelScenarioId;
+      const indications = metadata.IndicationId;
+      if (scenarios && indications) {
+        let file = await this.getFileWithSameTags(folder, scenarios, indications);
         if (file) this.appData.deleteFile(file?.ServerRelativeUrl);
       }
     }
@@ -130,13 +131,25 @@ export class FilesService {
     return commentsStr;   
   }
 
-  /** Search for a model with scenarios assigned */
-  async getFileByScenarios(path: string, scenarios: number[]) {
+  async firstCommentString(str: string) {
+    let currentUser = await this.appData.getCurrentUserInfo();
+    let newComment = {
+      text: str,
+      email: currentUser.Email,
+      name: currentUser.Title?.indexOf("@") == -1 ? currentUser.Title : currentUser.Email,
+      userId: currentUser.Id,
+      createdAt: new Date().toISOString()
+    }
+    return JSON.stringify([newComment]);
+  }
+
+  /** Search for a model with Scenarios and Indications assigned */
+  async getFileWithSameTags(path: string, scenarios: number[], indications: number[]): Promise<NPPFile | null> {
     let files = await this.appData.getFolderFiles(path, false);
     for (let i = 0; i < files.length; i++) {
       let model = files[i];
-      let sameScenario = this.haveSameScenarios(model, scenarios);
-      if (sameScenario) {
+      const sameTags = this.haveSameScenarios(model, scenarios) && this.haveSameIndications(model, indications);
+      if (sameTags) {
         return model;
       }
     }
@@ -183,21 +196,29 @@ export class FilesService {
       let data = { ApprovalStatusId: statusId };
       if (comments) Object.assign(data, { Comments: comments });
 
-      await this.appData.updateFilePropertiesById(file.ListItemAllFields.ID, rootFolder, data);
-      let res;
-      if (status === "Approved" && entity && file.ServerRelativeUrl.indexOf(FILES_FOLDER) == -1) {
-        let arrFolder = file.ServerRelativeUrl.split("/");
-        await this.removeOldApprovedModel(entity, file);
-        res = await this.appData.copyFile(file.ServerRelativeUrl, '/' + arrFolder[1] + '/' + arrFolder[2] + '/' + FOLDER_APPROVED + '/' + entity.BusinessUnitId + '/' + entity.ID + '/0/0/' + arrFolder[arrFolder.length - 3] + '/0/', file.Name);
+      try {
+        const updateResult = await this.appData.updateFilePropertiesById(file.ListItemAllFields.ID, rootFolder, data);
+        let newPath: string;
+        if (updateResult && status === "Approved" && entity && file.ServerRelativeUrl.indexOf(FILES_FOLDER) == -1) {
+          // copy the model to approved folder
+          let arrFolder = file.ServerRelativeUrl.split("/");
+          await this.removeOldApprovedModel(entity, file);
+          newPath = await this.appData.copyFile(file.ServerRelativeUrl, '/' + arrFolder[1] + '/' + arrFolder[2] + '/' + FOLDER_APPROVED + '/' + entity.BusinessUnitId + '/' + entity.ID + '/0/0/' + arrFolder[arrFolder.length - 3] + '/0/', file.Name);
+  
+          if (newPath) {
+            await this.appData.updateFilePropertiesByPath(newPath, {
+              OriginalModelId: file.ListItemAllFields.ID,
+              ApprovalDate: new Date().toISOString()
+            });
+            await this.copyCSV(file, newPath);
+          }
+          return !!newPath;
+        };
+        return updateResult;
 
-        if (res) {
-          await this.appData.updateFilePropertiesByPath(res, { OriginalModelId: file.ListItemAllFields.ID })
-          await this.copyCSV(file, res);
-        }
-        return res;
-      };
-
-      return true;
+      } catch(e) {
+        return false;
+      }
     } else {
       throw new Error("Missing file metadata.");
     }
@@ -281,12 +302,14 @@ export class FilesService {
   }
 
   private async removeOldApprovedModel(entity: Opportunity, file: NPPFile) {
-    if (file.ListItemAllFields && file.ListItemAllFields.ModelScenarioId) {
+    const scenarios = file.ListItemAllFields?.ModelScenarioId;
+    const indications = file.ListItemAllFields?.IndicationId;
+
+    if (file.ListItemAllFields && scenarios && indications) {
       const arrFolder = file.ServerRelativeUrl.split("/");
       const path = '/' + arrFolder[1] + '/' + arrFolder[2] + '/' + FOLDER_APPROVED + '/' + entity.BusinessUnitId + '/' + entity.ID + '/0/0/' + arrFolder[arrFolder.length - 3] + '/0/';
-      const scenarios = file.ListItemAllFields.ModelScenarioId;
 
-      const model = await this.getFileByScenarios(path, scenarios);
+      const model = await this.getFileWithSameTags(path, scenarios, indications);
       if (model) {
         await this.deleteFile(model.ServerRelativeUrl);
       }
@@ -307,6 +330,20 @@ export class FilesService {
       return sameScenario;
 
     } else return false;
+  }
+
+  /** Check if the model have exactly the same indications */
+  private haveSameIndications(model: NPPFile, indications: number[]): boolean {
+    if (model.ListItemAllFields && model.ListItemAllFields.IndicationId) {
+
+      let sameIndications = model.ListItemAllFields.IndicationId.length === indications.length;
+      for (let j = 0; sameIndications && j < model.ListItemAllFields.IndicationId.length; j++) {
+        let indicationId = model.ListItemAllFields.IndicationId[j];
+        sameIndications = sameIndications && (indications.indexOf(indicationId) != -1);
+      }
+      return sameIndications;
+    }
+    return false;
   }
 
   private async copyCSV(file: NPPFile, path: string) {
@@ -350,10 +387,7 @@ export class FilesService {
     let files: NPPFile[] = []
 
     if (powerBiLibrary && file.ListItemAllFields) {
-      const result = await this.appData.getFileByForecast(powerBiLibrary, file.ListItemAllFields.ID);
-      if (result.value) {
-        files = result.value;
-      }   
+      files = await this.appData.getFilesByForecast(powerBiLibrary, file.ListItemAllFields.ID);
     }
 
     return files;
@@ -361,15 +395,17 @@ export class FilesService {
 
   /** Clone a forecast model to a new file with new scenarios */
   async cloneForecastModel(originFile: NPPFile, newFilename: string, newScenarios: number[], authorId: number, comments = ''): Promise<boolean> {
-
     const destinationFolder = originFile.ServerRelativeUrl.replace('/' + originFile.Name, '/');
+
+    let fileWithSameTags = await this.getFileWithSameTags(destinationFolder, newScenarios, originFile.ListItemAllFields!.IndicationId);
+    if (fileWithSameTags) this.appData.deleteFile(fileWithSameTags.ServerRelativeUrl);
 
     let success = await this.appData.cloneFile(originFile.ServerRelativeUrl, destinationFolder, newFilename);
     if (!success) return false;
 
     let newFileInfo = await this.appData.getFileByName(destinationFolder, newFilename);
 
-    if (newFileInfo.value[0].ListItemAllFields && originFile.ListItemAllFields) {
+    if (newFileInfo[0].ListItemAllFields && originFile.ListItemAllFields) {
       const newData:any = {
         ModelScenarioId: newScenarios,
         Comments: comments ? comments : null,
@@ -379,9 +415,9 @@ export class FilesService {
       let arrFolder = destinationFolder.split("/");
       let rootFolder = arrFolder[3];
       
-      success = await this.appData.updateFilePropertiesById(newFileInfo.value[0].ListItemAllFields.ID, rootFolder, newData);
+      success = await this.appData.updateFilePropertiesById(newFileInfo[0].ListItemAllFields.ID, rootFolder, newData);
       if (success) {
-        await this.appData.changeFileEditor(authorId, rootFolder, newFileInfo.value[0].ListItemAllFields.ID);
+        await this.appData.changeFileEditor(authorId, rootFolder, newFileInfo[0].ListItemAllFields.ID);
       }
     }
     return success;
